General Framework Setup
A lot of testing back and forth.

Integration of both Facebook- and Google login in place.
ILoginMethod is an interface both shall inherit from to having a bloated login screen activity.

Dagger is setup:
HelloWorld-classes, components and modules are at the moment present in the project, but will be removed at a later date.

User:
A Model to represent both the logged in user, but also other chat members.