Expanding the framework further
- ChatMenuActivity added:
This menu will display the chatrooms. Will probably be using the MVVM to this as well as the future chat screen.

- Added SplashScreenActivity
Is displayed shortly until the application is loaded, and it's known if the user has logged in before. Then the user is either forwarded to the login screen, else to the ChatMenuActivity.

- Room is added to provide cache for the latest user, who has logged in. Its classes can be found in the room-package.

- RoomModule added:
Provides the functionality of Room.

- AppModule added:
Provides application context for Room, in order to keep Room separated from the activities.

- ImageManager added:
Communication to the Firebase Storage. I.e. image uploaded. Will probably be renamed in the future as it is a bit ambiguous.

- First Activity displayed is changed to SplashScreenActivity.