Sending Text Messages and beginning of adapter
- ChatMessagesAdapter added:
Not implemented yet. Going to be multipurpose in form of both displaying text messages as well as image messages.

- IMessage added:
All message types are going to inherit from this. Makes it easier to display in the future RecyclerView.

- TextMessage added:
Standard text message model. Must receive a userId and String to be able to send.