Deletion and addition of load amount
- Deletions:
The different unused classes and interface were deleted. Including the HelloWorld-classes used for testing.

- IMessage changed to Message:
It is now an abstract class, because of the shared functionality between ImageMessage and TextMessage.

- ActiveChatManager modified:
It can now fetch a range of messages. Used in the future for the displayment. Both supports new messages as well as a range of them.